Refactor dependencies
