+++ ReflectometryServer/test_modules/test_configuration.py
@@ -2,8 +2,7 @@ import os
 import unittest
 
 from ReflectometryServer import beamline_configuration
-from ReflectometryServer.beamline import Beamline
-from ReflectometryServer.ChannelAccess.pv_manager import STATUS
+from ReflectometryServer.beamline import Beamline, STATUS
 
 
 class TestConfiguration(unittest.TestCase):
